add persistent URL param support to visit, request_page
This patch adds a fourth argument 'persistent_data' to
Webrat::Session#visit and Webrat::Session#request_page in order
to support request parameters that should be persistently applied
along a chain of redirects.  That is, if request_page follows
a redirect, it supplies the parameters in 'persistent_data' to
the redirected request, and passes them along to the next request
as well.

My use case for this is an application whose caller is a portal
system that communicates information specific to each request as
URL parameters.  In order to simulate being called by the portal
system in testing, these parameters need to persist through
redirects Webrat follows.